LayoutTests/media/media-can-play-flac-audio.html

<!doctype html>
<html>
    <head>
        <script src=video-test.js></script>
        <script>
            function start()
            {
                audio = document.getElementsByTagName('audio')[0];

                // FLAC encoded audio in a native flac container
                testExpected("audio.canPlayType('audio/flac')", "maybe");
                testExpected("audio.canPlayType('audio/x-flac')", "maybe");

                endTest();
            }
        </script>

    </head>
    <body onload="start()">

        <audio controls ></audio>

        <p>Test HTMLMediaElement <em>canPlayType()</em> method with multiple .flac MIME types.</p>

    </body>
</html>
