LayoutTests/media/media-can-play-flac-audio-expected.txt

Test HTMLMediaElement canPlayType() method with multiple .flac MIME types.

EXPECTED (audio.canPlayType('audio/flac') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/x-flac') == 'maybe') OK
END OF TEST
